fix: 🐛 #548 s init流程 针对yaml格式问题请提示具体些 (参考s verify)
