+++ src/libs/index.ts
@@ -1,3 +1,4 @@
 export * from './utils';
 export * from './common';
 export { default as getYamlContent } from './getYamlContent';
+export { default as isYamlFile } from './isYamlFile';
